output/ssh: generic tx json logger
